휴대형구름 기능 개선 및 오류 수정 - Jenkins 빌드 서버 연동 파라미터 추가 (noti_email, user_pw, user_id) - DB 쿼리 개선 (이미지 리스트 검색 오류) - 이미지 사이즈 타입 변경 (int -> long)

+++ src/main/java/kr/gooroom/gpms/ptgr/service/PortableImageVO.java
@@ -8,7 +8,7 @@ import java.util.Date;
 public class PortableImageVO implements Serializable {
 
     int imageId;
-    int size;
+    long size;
     String name;
     String url;
     int status; //신청/생성/경로전송/완료/삭제
@@ -23,11 +23,11 @@ public class PortableImageVO implements Serializable {
         this.imageId = imageId;
     }
 
-    public int getSize() {
+    public long getSize() {
         return size;
     }
 
-    public void setSize(int size) {
+    public void setSize(long size) {
         this.size = size;
     }
